agent-platform-repos/memory-rag-service/pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "memory-rag-service"
version = "0.1.0"
requires-python = ">=3.11"
dependencies = [
    "agent-platform-shared @ file://../shared-lib",
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.30.0",
    "pymilvus>=2.4.0",
    "qdrant-client>=1.11.0",
    "sentence-transformers>=3.0.0",
    "FlagEmbedding>=1.2.0",
    "mem0ai>=0.1.0",
    "redis>=5.1.0",
    "httpx>=0.27.0",
    "tenacity>=9.0.0",
]

[project.optional-dependencies]
dev = ["pytest>=8.0.0", "pytest-asyncio>=0.24.0", "pytest-mock>=3.14.0", "ruff>=0.7.0"]

[tool.pytest.ini_options]
asyncio_mode = "auto"
testpaths = ["tests"]
